==> 05_logit_2022.Rmd <==
---
title: "Phylogenetic knowledge vs. range size"
author: "Wolf L. Eiserhardt"
date: "7/7/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
```{r}
R.Version()$version.string
```

## Logistic regression

Logistic regression of the sequencing status of species (1 = sequenced, 0 = not sequenced) on their range size, measured as the sum of the areas (in km^2) of all the botanical countries a species occurs in. 

Load required libraries:
```{r, message = FALSE}
library(data.table)
library(ggplot2)
library(MASS)
library(viridis)
```

Convenience function for density plot (from https://slowkow.com/notes/ggplot2-color-by-density/):
```{r}
get_density <- function(x, y, ...) {
  dens <- MASS::kde2d(x, y, ...)
  ix <- findInterval(x, dens$x)
  iy <- findInterval(y, dens$y)
  ii <- cbind(ix, iy)
  return(dens$z[ii])
}
```

Load data:
```{r}
library("rstudioapi")
setwd(dirname(getActiveDocumentContext()$path))
dat <- fread("data\\seq_area_2022.csv")
dat <- as.data.frame(dat)
```

Fit logistic regression:
```{r}
glm.fit <- glm(sequenced ~ area_sq, data = dat, family = binomial)
summary(glm.fit)
```

Prepare data for plotting:
```{r}
# Fitted probabilities
preddata <- data.frame(area_sq=seq(min(dat$area_sq), max(dat$area_sq),len=500))
preddata$sequenced = predict(glm.fit, preddata, type="response")
# Point density
dat$density <- get_density(dat$area_sq, dat$sequenced, n = 100)
```

Plot model:
```{r}
tiff(filename="data\\logit_reg_HR.tiff",height=5600,width=5200,units= 'px' ,res=800,compression = 'lzw' )
ggplot(dat) + geom_point(aes(area_sq, sequenced, color = density)) + scale_color_viridis() + geom_line(data=preddata,aes(x=area_sq,y=sequenced), color = "red", size=1) + labs(x="range size")
dev.off()
```
